requirements.txt
pygame-ce
pygame-menu-ce
requests
datetime
pygamevideo